Code adjustment
